+++ src/main/java/com/petchatbot/repository/PetRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PetRepository extends JpaRepository<Pet, Integer> {
 
-    Pet findByPetSerial(Long petSerial);
+    Pet findByPetSerial(int petSerial);
 
     Member findByMember(Member member);
